use ints instead of doubles for ratings now

+++ src/main/sc/player2023/logic/pvs/MoveGetterUtil.java
@@ -16,7 +16,7 @@ public final class MoveGetterUtil {
     }
 
 
-    static double getRatingFactorForNextMove(@Nonnull ImmutableGameState gameState) {
+    static int getRatingFactorForNextMove(@Nonnull ImmutableGameState gameState) {
         if(GameRuleLogic.anyPossibleMovesForPlayer(gameState.getBoard(), gameState.getCurrentTeam().opponent())) {
             return -1;
         }
